看板可视编辑：删除元素、解绑图表操作改为需确认执行

+++ datagear-web/src/main/resources/org/datagear/web/templates/dashboard/include/dashboard_form_editor.ftl
@@ -1402,11 +1402,23 @@
 				{
 					if(deleteOperation == "deleteElement")
 					{
-						dashboardEditor.deleteElement();
+						po.confirm("<@spring.message code='dashboard.opt.delete.element.confirm' />",
+						{
+							confirm: function()
+							{
+								dashboardEditor.deleteElement();
+							}
+						});
 					}
 					else if(deleteOperation == "unbindChart")
 					{
-						dashboardEditor.unbindChart();
+						po.confirm("<@spring.message code='dashboard.opt.delete.unbindChart.confirm' />",
+						{
+							confirm: function()
+							{
+								dashboardEditor.unbindChart();
+							}
+						});
 					}
 				}
 			}
